CRGBA class update
Now it is possible to create two types of CRGBA: one points to a memory address and it is used by the game, the other is a simple Python class instance which can be pretty useful in scripts as you can create your own instances using the static functions with 'New' prefix.
This concept will be hopefully extended to every exposed class, to grant a widest range of opportunities in your scripts.

Keep in mind that game classes, as works with preallocated memory and use game functions, are usually faster than python instances.